requirements.txt
pycryptodome
pillow
pytest